+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Injecting sugary liquids will directly increase blood sugar
 - Electronic watches periodically report your blood sugar
 - Metformin and glucagon containers now drop to 0.1u when empty, not 0u
+- New Blood glucose meter item can be used to manually check your blood sugar when unchipped
+- Blood glucose meter is now needed to craft a Smart insulin pump
 
 ## 1.1.0
 - Game no longer crashes when Casualties: Together isn't installed